# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QuickLink API",
        "description": "API for QuickLink healthcare platform",
        "contact": {
            "email": "support@quicklink.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "QuickLink API Server"
        }
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User registration",
                "description": "Register a new user account. Supports patient (default) and doctor role. If role=doctor, doctor fields must be provided.",
                "operationId": "c7bb21a0d87049e161b8e4c9b0d909f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "password",
                                    "dateOfBirth",
                                    "gender",
                                    "address",
                                    "emergencyContact",
                                    "emergencyContactPhone",
                                    "bloodGroup"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+255700000000"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "securePassword123"
                                    },
                                    "dateOfBirth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Mikocheni, Dar es Salaam"
                                    },
                                    "emergencyContact": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "emergencyContactPhone": {
                                        "type": "string",
                                        "example": "+255700000001"
                                    },
                                    "medicalHistory": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "diabetes",
                                            "hypertension"
                                        ]
                                    },
                                    "allergies": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "penicillin"
                                        ]
                                    },
                                    "bloodGroup": {
                                        "type": "string",
                                        "example": "O+"
                                    },
                                    "profileImage": {
                                        "description": "Profile image file (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "role": {
                                        "description": "Defaults to patient if omitted",
                                        "type": "string",
                                        "enum": [
                                            "patient",
                                            "doctor"
                                        ],
                                        "example": "patient"
                                    },
                                    "hospitalId": {
                                        "description": "Required if role=doctor (or first hospital will be used)",
                                        "type": "string",
                                        "example": "1",
                                        "nullable": true
                                    },
                                    "licenseNumber": {
                                        "description": "Required if role=doctor",
                                        "type": "string",
                                        "example": "TMC-123456",
                                        "nullable": true
                                    },
                                    "specialty": {
                                        "description": "Required if role=doctor",
                                        "type": "string",
                                        "example": "Cardiology",
                                        "nullable": true
                                    },
                                    "yearsOfExperience": {
                                        "description": "Required if role=doctor",
                                        "type": "integer",
                                        "example": 8,
                                        "nullable": true
                                    },
                                    "clinicName": {
                                        "type": "string",
                                        "example": "Sunrise Clinic",
                                        "nullable": true
                                    },
                                    "clinicAddress": {
                                        "type": "string",
                                        "example": "123 Main St, Dar es Salaam",
                                        "nullable": true
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Cardiologist with a focus on preventive care",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "user_123"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+255700000000"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "patient"
                                                },
                                                "profileImageUrl": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/storage/profile_images/user_123.jpg"
                                                },
                                                "createdAt": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-15T10:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "jwt_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "description": "Logout user and invalidate JWT token",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout successful"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}